Lakeshore_plugin/pyproject.toml
[build-system]
requires = [
    "setuptools>=61.0.0",
    "setuptools-scm>=8.0",
]
build-backend = "setuptools.build_meta"

[project]
name = "lakeshore_plugin"
version = "0.0.1"
description = "A plugin for NOMAD containing Lakeshore hall measurements."
readme = "README.md"
authors = [
    { name = "Hampus Näsström", email = 'hampus.naesstroem@physik.hu-berlin.de' },
    { name = "Andrea Albino", email = 'andrea.albino@physik.hu-berlin.de' },
    { name = "Sarthak Kapoor", email = 'sarthak.kapoor@ikz-berlin.de' },
    { name = "Sebastian Brückner", email = 'sebastian.brueckner@ikz-berlin.de' },
]
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "License :: OSI Approved :: Apache Software License",
]
dependencies = [
    "nomad-lab>=1.2.1",
]
[project.optional-dependencies]
dev = [
    "pytest",
    "structlog>=22.3.0",
]

[project.license]
file = "LICENSE"

[project.urls]
"Homepage" = "https://github.com/FAIRmat-NFDI/nomad-material-processing"
"Bug Tracker" = "https://github.com/FAIRmat-NFDI/nomad-material-processing/issues"

[tool.setuptools.packages.find]
where = [
    "src",
]

[project.entry-points.'nomad.plugin']

hall_schema = "lakeshore_plugin.hall:hall_schema"
hall_instrument_parser = "lakeshore_plugin.hall.instrument_parser:hall_instrument_parser"
hall_measurement_parser = "lakeshore_plugin.hall.measurement_parser:hall_measurement_parser"

